requirements.txt
asgiref==3.5.2
astroid==2.12.12
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
cryptography==37.0.4
dill==0.3.6
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.2
django-filter==22.1
django-heroku==0.0.0
djangorestframework==3.14.0
dulwich==0.20.46
Faker==15.1.1
filelock==3.8.0
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
idna==3.4
isort==5.10.1
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lazy-object-proxy==1.8.0
lockfile==0.12.2
Markdown==3.4.1
mccabe==0.7.0
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
pipenv==2022.10.12
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.1.2
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pylint==2.15.5
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==1.5
pytz==2022.5
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typed-ast==1.5.4
urllib3==1.26.12
validate-docbr==1.10.0
virtualenv==20.16.6
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.12.1
